pyproject.toml
[build-system]
requires = ["hatchling>=1.0.0"]
build-backend = "hatchling.build"

[project]
name = "gdankh"
version = "0.1.0"
description = "library for truly rust developers"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "maslina524"}
]
maintainers = [
    {name = "maslina524"}
]
keywords = ["rust", "modern"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]

dependencies = [
    "websocket-client"
]

[project.optional-dependencies]
dev = []
test = []
docs = []


[project.urls]
Homepage = "https://github.com/maslina524/gdankh"
# Documentation = "https://my-awesome-package.readthedocs.io"
# Repository = "https://github.com/username/my-awesome-package.git"
# Issues = "https://github.com/username/my-awesome-package/issues"

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I", "N", "UP"]

[tool.mypy]
python_version = "3.9"
strict = true